Login funcionando

+++ ecommerce-f1/app/Http/Controllers/ProdutoController.php
@@ -10,20 +10,29 @@ class ProdutoController extends Controller
 {
     public function index(){
         $produtos = Produto::all();//retorna todos os produtos e guarda nessa variavel
+        $usuario = $this->quemEstaLogado();
+        if (!$usuario) return redirect('/user/login');
 
-      return view('produto.index')->with('produtos',$produtos);//retorna a view numa pasta(n pode ser no plural, pois é o q está na model) e o arquivo("".blade.php)
+      return view('produto.index')->with('produtos',$produtos)->with('usuario', $usuario);//retorna a view numa pasta(n pode ser no plural, pois é o q está na model) e o arquivo("".blade.php)
     }
 
     public function show(Produto $produto){ // model e variavel
 
       $categorias = []; //Categoria::find($produto->CATEGORIA_ID)->Produtos;
-      return view ('produto.show', ['produto' =>$produto,'categorias' => $categorias]);
+      $usuario = $this->quemEstaLogado();
+      if (!$usuario) return redirect('/user/login');
+      return view ('produto.show', ['produto' =>$produto,'categorias' => $categorias, 'usuario' => $usuario]);
 
     }
     public function create(){
-      return view ('produto.create', ['categorias' => []]);
+      $usuario = $this->quemEstaLogado();
+      if (!$usuario) return redirect('/user/login');
+      return view ('produto.create', ['usuario' => $usuario, 'categorias' => []]);
     }
     public function store(Request $request){
+      $usuario = $this->quemEstaLogado();
+      if (!$usuario) return redirect('/user/login');
+
       $produto = new Produto();
       $produto->nome = $request->nome;
       $produto->descricao = $request->descricao;
@@ -39,9 +48,14 @@ class ProdutoController extends Controller
     }
 
     public function edit (Produto $produto){
-      return view ('produto.edit', ['produto' => $produto,'categorias' => []]);
+      $usuario = $this->quemEstaLogado();
+      if (!$usuario) return redirect('/user/login');
+      return view ('produto.edit', ['usuario' => $usuario, 'produto' => $produto,'categorias' => []]);
     }
     public function update(Request $request, Produto $produto){
+      $usuario = $this->quemEstaLogado();
+      if (!$usuario) return redirect('/user/login');
+
       $produto->nome = $request->nome;
       $produto->descricao = $request->descricao;
       $produto->tamanho = $request->tamanho;
@@ -56,6 +70,8 @@ class ProdutoController extends Controller
     }
 
     public function destroy (Produto $produto){
+      $usuario = $this->quemEstaLogado();
+      if (!$usuario) return redirect('/user/login');
       $produto->delete();
       return redirect ('produto');
     }
